Move more out of app/core
