+++ apps/zui/src/ppl/detail/UidPanel.tsx
@@ -6,7 +6,7 @@ import PanelHeading from "src/views/detail-pane/PanelHeading"
 import EventTimeline from "src/ppl/detail/EventTimeline"
 import {SecurityEvent} from "src/ppl/detail/models/security-event"
 import {sort} from "src/ppl/detail/util/sort"
-import {Data, Name, Value} from "src/app/core/Data"
+import {Data, Name, Value} from "src/components/data"
 import formatDur from "src/ppl/detail/util/formatDur"
 import {isEqual} from "lodash"
 import Panel from "src/views/detail-pane/Panel"
