+++ apps/zui/src/test/system/boot.tsx
@@ -3,7 +3,7 @@ import React from "react"
 import {main} from "src/electron/run-main/run-main"
 import initialize from "src/js/initializers/initialize"
 import fsExtra from "fs-extra"
-import {AppProvider} from "src/app/core/context"
+import {AppProvider} from "src/views/application/context"
 import {getPort} from "./port-service"
 import {waitFor} from "@testing-library/react"
 import {Store} from "src/js/state/types"
