+++ apps/zui/src/views/session-page/use-title-form.ts
@@ -1,6 +1,6 @@
 import {FormEvent} from "react"
 import {useSelector} from "react-redux"
-import {useZuiApi} from "src/app/core/context"
+import {useZuiApi} from "src/views/application/context"
 import {useDispatch} from "src/app/core/state"
 import Layout from "src/js/state/Layout"
 import Current from "src/js/state/Current"
