+++ apps/zui/src/views/sidebar/pools-section/pools-tree.tsx
@@ -11,7 +11,7 @@ import {Empty} from "./empty"
 import {FillFlexParent} from "src/components/fill-flex-parent"
 import PoolItem from "./pool-item"
 import {groupByDelimiter, Internal} from "./group-by"
-import {useZuiApi} from "src/app/core/context"
+import {useZuiApi} from "src/views/application/context"
 import * as poolCmd from "src/app/commands/pools"
 import Config from "src/js/state/Config"
 import {Pool} from "src/app/core/pools/pool"
