+++ apps/zui/src/views/sidebar/queries-section/queries-tree.tsx
@@ -2,7 +2,7 @@ import React, {useRef} from "react"
 import {Tree, TreeApi} from "react-arborist"
 import {useSelector} from "react-redux"
 import {deleteQueries} from "src/app/commands/delete-queries"
-import {useZuiApi} from "src/app/core/context"
+import {useZuiApi} from "src/views/application/context"
 import {useDispatch} from "src/app/core/state"
 import {queryTreeContextMenu} from "src/app/menus/query-tree-context-menu"
 import Current from "src/js/state/Current"
@@ -12,7 +12,7 @@ import {DropOverlay} from "../drop-overlay"
 import {useQueryImportOnDrop} from "../hooks"
 import {FillFlexParent} from "src/components/fill-flex-parent"
 import QueryItem from "./query-item"
-import {selectQuery} from "src/app/events/select-query-event"
+import {selectQuery} from "src/views/sidebar/queries-section/select-query-event"
 import Appearance from "src/js/state/Appearance"
 import {TREE_ITEM_HEIGHT} from "../item"
 import {showMenu} from "src/core/menu"
