+++ apps/zui/src/views/sidebar/queries-section/select-query-event.ts
@@ -1,3 +1,3 @@
-import {createEvent} from "../core/events"
+import {createEvent} from "./events"
 
 export const selectQuery = createEvent<[id: string]>("selectQuery")
